+++ src/index.ts
@@ -1,6 +1,5 @@
 import Router from './router/router';
 import Config from './config';
-import Component from './component/component';
 import WebComponent from './component/web_component';
 
 export * as i18n from './intl/i18n';
@@ -10,4 +9,4 @@ export { html } from './template/template';
 export { createState } from './state/state';
 export { registerCss } from './css';
 export { defineWebComponent } from './component/web_component';
-export { Router, Config, Component, WebComponent };
+export { Router, Config, WebComponent };